+++ inst/examples/app.R
@@ -228,24 +228,24 @@ shinyApp(
             width = 6,
             descriptionBlock(
               number = "17%", 
-              number_color = "green", 
-              number_icon = "fa fa-caret-up",
+              numberColor = "green", 
+              numberIcon = "fa fa-caret-up",
               header = "$35,210.43", 
               text = "TOTAL REVENUE", 
-              right_border = TRUE,
-              margin_bottom = FALSE
+              rightBorder = TRUE,
+              marginBottom = FALSE
             )
           ),
           dashboardUserItem(
             width = 6,
             descriptionBlock(
               number = "18%", 
-              number_color = "red", 
-              number_icon = "fa fa-caret-down",
+              numberColor = "red", 
+              numberIcon = "fa fa-caret-down",
               header = "1200", 
               text = "GOAL COMPLETION", 
-              right_border = FALSE,
-              margin_bottom = FALSE
+              rightBorder = FALSE,
+              marginBottom = FALSE
             )
           )
         )