+++ inst/examples/box_elements/description_block.R
@@ -10,7 +10,7 @@ description_block <- 'box(
       descriptionBlock(
         number = "17%", 
         numberColor = "green", 
-        numberIcon = "fa fa-caret-up",
+        numberIcon = "caret-up",
         header = "$35,210.43", 
         text = "TOTAL REVENUE", 
         rightBorder = TRUE,
@@ -22,7 +22,7 @@ description_block <- 'box(
       descriptionBlock(
         number = "18%", 
         numberColor = "red", 
-        numberIcon = "fa fa-caret-down",
+        numberIcon = "caret-down",
         header = "1200", 
         text = "GOAL COMPLETION", 
         rightBorder = FALSE,